requirements.txt
praw
psycopg2
pyyaml
